Added interface deconfliction for classes. This removes the need to removes the need to manually edit metadata files when there are conflicts. Eg: IMapControl::get_Style and IFrameworkElement::get_Style which are both implemented by MapControl.

+++ code-generator/src/main/kotlin/com/github/knk190001/winrtbinding/generator/model/entities/SparseInterface.kt
@@ -19,7 +19,7 @@ data class SparseInterface(
     @Json("Traits")
     val traits: List<Trait>,
     @Json("ContractVersion")
-    val contractVersion: UInt?
+    val contractVersion: Long?
 ): SparseEntity("Interface"), IDirectProjectable<SparseInterface>{
     override fun projectType(typeVariable: String, newTypeReference: SparseTypeReference): SparseInterface {
         return this.copy(
